Added fifth grid pass (enemies and boss)
Enemy design, generation and placement finished. Grid is mapped again and enemies have different spawn probability depending on which third of the map is being mapped. Easier enemies more likely on bottom, average in the middle, hard up top.

Finally, a new map pass is made and the first of the hard enemies found in the grid is replaced by the boss, ensuring it is placed at one of the highest possible valid positions.